+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Upcoming release
+
+- Expose Notificare base methods and properties statically to the JVM
+- Add intent actions & extras as JVM fields
+
 ## 3.3.0
 
 - Add opt-in intent when opening a beacon foreground service notification